Fixed typos with Weight Page section
Great checklist! Also changed the Page Weight tool to a more user friendly and less ad covered tool.
Also wondering in the Font Sections it mentions ```2 MO``` is this meant to be ```2 MB``` or something else? A quick google search gave me nothing about MO so a little explanation would be great!